Minor updates
Made this guild-only
Updated small formatting
Changed ascii characters into emojis and made them random!
