+++ src/commands/guild/lovebomb.js
@@ -1,0 +1,25 @@
+'use strict';
+
+const path = require('path');
+const { SlashCommandBuilder } = require('@discordjs/builders');
+const logger = require('../../utils/logger');
+
+const PREFIX = path.parse(__filename).name;
+
+const heartEmojis = [
+  '❤', '🧡', '💛', '💚', '💙', '💜',
+  '💝', '💖', '💗', '💘', '💕', '💞', '💓', '💟', '❣',
+];
+
+module.exports = {
+  data: new SlashCommandBuilder()
+    .setName('lovebomb')
+    .setDescription('Spread some love'),
+
+  async execute(interaction) {
+    const message = `${heartEmojis.sort(() => 0.5 - Math.random()).slice(0, 30).join(' ')}`;
+    interaction.reply(message);
+
+    logger.debug(`[${PREFIX}] finished!`);
+  },
+};
